use no-op models when we don't need real ones for summary functionality
