+++ llm-service/app/tests/conftest.py
@@ -42,7 +42,7 @@ import uuid
 from collections.abc import Iterator
 from dataclasses import dataclass
 from datetime import datetime
-from typing import Any, Dict, Sequence
+from typing import Any, Dict
 
 import boto3
 import lipsum
@@ -51,24 +51,14 @@ import qdrant_client as q_client
 from boto3.resources.base import ServiceResource
 from fastapi.testclient import TestClient
 from llama_index.core.base.embeddings.base import BaseEmbedding, Embedding
-from llama_index.core.base.llms.types import (
-    ChatMessage,
-    ChatResponse,
-    ChatResponseAsyncGen,
-    ChatResponseGen,
-    CompletionResponse,
-    CompletionResponseAsyncGen,
-    CompletionResponseGen,
-    LLMMetadata,
-)
 from llama_index.core.llms import LLM
 from moto import mock_aws
-from pydantic import Field
 
 from app.ai.vector_stores.qdrant import QdrantVectorStore
 from app.main import app
 from app.services import models, data_sources_metadata_api
 from app.services.data_sources_metadata_api import RagDataSource
+from app.services.noop_models import DummyLlm
 from app.services.utils import get_last_segment
 
 
@@ -129,62 +119,6 @@ def index_document_request_body(
             "chunk_overlap": 10,
         },
     }
-
-
-class DummyLlm(LLM):
-    completion_response: str = Field("this is a completion response")
-    chat_response: str = Field("this is a chat response")
-
-    def __init__(
-        self,
-        completion_response: str = "this is a completion response",
-        chat_response: str = "hello",
-    ):
-        super().__init__()
-        self.completion_response = completion_response
-        self.chat_response = chat_response
-
-    @property
-    def metadata(self) -> LLMMetadata:
-        return LLMMetadata()
-
-    def chat(self, messages: Sequence[ChatMessage], **kwargs: Any) -> ChatResponse:
-        return ChatResponse(message=ChatMessage.from_str(self.chat_response))
-
-    def complete(
-        self, prompt: str, formatted: bool = False, **kwargs: Any
-    ) -> CompletionResponse:
-        return CompletionResponse(text=self.completion_response)
-
-    def stream_chat(
-        self, messages: Sequence[ChatMessage], **kwargs: Any
-    ) -> ChatResponseGen:
-        raise NotImplementedError("Not implemented")
-
-    def stream_complete(
-        self, prompt: str, formatted: bool = False, **kwargs: Any
-    ) -> CompletionResponseGen:
-        raise NotImplementedError("Not implemented")
-
-    async def achat(
-        self, messages: Sequence[ChatMessage], **kwargs: Any
-    ) -> ChatResponse:
-        raise NotImplementedError("Not implemented")
-
-    async def acomplete(
-        self, prompt: str, formatted: bool = False, **kwargs: Any
-    ) -> CompletionResponse:
-        raise NotImplementedError("Not implemented")
-
-    async def astream_chat(
-        self, messages: Sequence[ChatMessage], **kwargs: Any
-    ) -> ChatResponseAsyncGen:
-        raise NotImplementedError("Not implemented")
-
-    async def astream_complete(
-        self, prompt: str, formatted: bool = False, **kwargs: Any
-    ) -> CompletionResponseAsyncGen:
-        raise NotImplementedError("Not implemented")
 
 
 class DummyEmbeddingModel(BaseEmbedding):
